Fix bug in DIC_NR, move file writing functions to test (for now)
